Add unit tests for KeyserverHttpClient
This should be extracted so it can be reused in other projects in
order to test the various service clients we have, which currently
have no unit tests.

This allows us to unit test an http client against the akka routes
defined in the project, which are exactly the routes that are offered
by the akka service.

Users of this client should update their implementations to use the
companion object factory method instead of calling `new` directly.