Add --no-auth flag to `init` command
When a repository is initialized with `--no-auth`, subsequent commands
will not contact auth+ to obtain a token before contacting tuf repo.

Import `tufrepo.url` file containing the tuf url when initializing a
repository from a `credentials.zip`.

Export currently configured `tufrepo.url` when exporting repository
credentials.zip.

Removed support for `--reposerver` flag when not using command
`init`. A repository can still be initialized to connect to a specific
tuf reposerver, but that url is set when initializing a repository and
cannot be changed through the comand line. However, this url can be
changed editing the `config.json` file in the repository
directory. `--reposerver` is mandatory when initializing a repository
through `treehub.json`.