Refresh role dependencies when role is refreshed
When snapshots.json was refreshed due to an expiration,
timestamps.json was not refreshed to reflect that change. The same was
happening with targets.json and it's dependencies (snapshots.json -> targets.json)

This required a sizeable refactor otherwise we could not have the
refresh operation with type safe signatures as didn't have a way
during compilation to distinguish from a targets.json in SignedRole or
a snapshots.json in a SignedRole, for example.