Return expired root.json if keys are offline
Without this change, a user cannot `garage-sign root pull` as
keyserver always returns an error since it tries to refresh the role,
but the keys to do so are offline.

With this change the expired root.json is returned and is up to the
client to validate it according to the tuf/update spec.

I don't like that I am quite explicitly using an exception for control
flow, but the other alternative would be to return a
`Future[ResultType]`, and then have a dedicated `ResultType` for this,
but I think it's too complex to use something like that.

Signed-off-by: Simão Mata <simao.mata@here.com>