Stream file upload to s3 if client supports it
If client uploads the file contents as request entity, and a valid
`Content-Length` header, upload file to s3 instead of buffering
locally.

This is equivalent to the following curl command:

```
curl -H 'application/octet-stream' --data-binary @myfile.bin host:/api/v2/objects/:object-i
```

Not this is different from the request the UI currently uses, which
includes a form field and does not provide a valid size for the file.