Support big files upload to tuf-reposerver when using s3
This includes both client (cli/garage-sign ) and server support (reposerver).

A new endpoint is implemented that accepts a request containing binary
data and redirects the client to s3 for out-of-band
uploading. However, the target is not immediately added to the user's
`targets.json`, in order to easily support offline targets keys.

The offline targets update endpoint was updated to check the `custom`
field on each target, and if it contains a specific attribute, the
target will be marked as uploaded out of band and a `GET` to this
target's filename will be redirected to S3 using the usual `Managed`
target logic.

Support in `garage-sign` for these endpoints was added in the form of
two new commands:

`garage-sign` targets upload` Will upload a given binary to
reposerver/s3 directly. Since this is done directly to s3, this
endpoint accepts big files, the limitation is in the s3
policy. However, the content-length will be limited by the pre-sign
url generated by reposerver.

After the upload to s3 is successful, the user can use new command
`garage-sign targets add-uploaded` to add the file to it's
targets. After signing, the new `targets.json` can be pushed to the
server. A new README section was added with the complete steps.

Refer to `garage-sign --help` to learn more about these commands.

To implement this functionality, we needed to use a different http
client in cli, as the client we were using does not support uploading
a file to the server without first putting the file in
memory.

The `--verbose` command in `garage-sign` will now make `garage-sign`
print a complete `String` representation of the request to `stdout`,
as well as a `curl` command that can be used to reproduce the request
on a CLI.

Signed-off-by: Simão Mata <simao.mata@here.com>