Parse the OAuth2 credentials using the format expected by Cognito.
When migrating Auth+ to AWS Cognito, there is a new `scope`
field in the `treehub.json`. That was implicitly managed by
Auth+, but it needs to be used explicitly for Cognito. So we
need to parse that new field from the `treehub.json` and use
it when generating a token.

This change is backwards-compatible, so current versions of
the credentials.zip can still be used to request tokens from
Auth+. This way clients can upgrade garage-sign in advance.
Then we can safely migrate Auth+ to Cognito without having to
synchronize with clients.

Relates-to: OTA-5298
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>