Adjust `targets sign` to sign given multiple keyName - signature pairs.
Up until now, `targets sign` accepts the keyId and the signature to add
to the targets role. From now on, it accepts a map of keyName and signature.
That means that we can add multiple signatures at once to the target role,
and that the command is arguably more user-friendly, since the user has
to pass the name of the file of the key they want to use, instead of the keyId.

Relates-to: OTA-5319
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>