Fix namespace validation.
During my previous commit 6d760146 I removed the last unused auth+ config. Apparently
the `NamespaceValidation` was making use of that under the hood and I didn't notice.
But as we were using `auth.protocol = none` and `auth.verification = none`, we can
simply explicitly use the `defaultNamespaceExtractor`. It's the one we've been implicitly
using up until now, as per the configuration we had.

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>